+++ lib/ts/recipe/emailverification/recipe.tsx
@@ -66,7 +66,7 @@ export default class EmailVerification extends RecipeModule<
     }
 
     static init(
-        config: UserInput
+        config?: UserInput
     ): RecipeInitResult<GetRedirectionURLContext, PreAndPostAPIHookAction, OnHandleEventContext, NormalisedConfig> {
         const normalisedConfig = normaliseEmailVerificationFeature(config);