+++ rollup.config.mjs
@@ -19,7 +19,7 @@ export default [
             "ui-entry": "lib/ts/ui/index.tsx",
             session: "lib/ts/recipe/session/index.ts",
             sessionprebuiltui: "lib/ts/recipe/session/prebuiltui.tsx",
-            emailverification: "lib/ts/recipe/emailverification/index.tsx",
+            emailverification: "lib/ts/recipe/emailverification/index.ts",
             emailverificationprebuiltui: "lib/ts/recipe/emailverification/prebuiltui.tsx",
             emailpassword: "lib/ts/recipe/emailpassword/index.ts",
             emailpasswordprebuiltui: "lib/ts/recipe/emailpassword/prebuiltui.tsx",